Introduce dns_random_engine and use it in shuffle
